Refactor game file structure
