+++ game/lib/components/player/bullet_component.dart
