+++ game/lib/components/player/player_component.dart
@@ -1,7 +1,7 @@
 import 'package:flame/components.dart';
 
-import 'package:game/rogalik_game.dart';
-import 'package:game/components/players/player.dart';
+import 'package:game/game/rogalik_game.dart';
+import 'package:game/entities/player/player.dart';
 import 'package:game/config/game_config.dart';
 
 class PlayerComponent extends SpriteComponent with HasGameRef<RogalikGame> {
