+++ game/lib/entities/common/entity_stats.dart
