+++ game/lib/entities/enemies/enemy.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 import 'package:flame/components.dart';
 
-import 'package:game/components/common/entity_stats.dart';
+import 'package:game/entities/common/entity_stats.dart';
 
 class Enemy {
   EntityStats stats;
