+++ game/lib/entities/enemies/enemy_factory.dart
@@ -1,6 +1,6 @@
-import 'package:game/components/enemies/enemy.dart';
-import 'package:game/components/common/entity_stats.dart';
-import 'package:game/components/enemies/types/zombie_enemy.dart';
+import 'package:game/entities/enemies/enemy.dart';
+import 'package:game/entities/common/entity_stats.dart';
+import 'package:game/entities/enemies/zombie_enemy.dart';
 
 class EnemyFactory {
   static Enemy createEnemy(String type, EntityStats stats) {
