+++ game/lib/entities/enemies/zombie_enemy.dart
@@ -1,7 +1,7 @@
 import 'package:flame/extensions.dart';
 
-import 'package:game/components/enemies/enemy.dart';
-import 'package:game/components/common/entity_stats.dart';
+import 'package:game/entities/enemies/enemy.dart';
+import 'package:game/entities/common/entity_stats.dart';
 
 class Zombie extends Enemy {
   Zombie(EntityStats stats) : super(stats);
