+++ game/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flame/game.dart';
 import 'package:flutter/widgets.dart';
-import 'package:game/rogalik_game.dart';
+import 'package:game/game/rogalik_game.dart';
 
 void main() {
   runApp(GameWidget(game: RogalikGame()));
